requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
tweepy>=4.14.0
google-genai>=1.0.0
pydantic>=2.5.0
pyyaml>=6.0.1
click>=8.1.0
apscheduler>=3.10.0
PyPDF2>=3.0.0
